fix(deps): find tools the app can see but PATH does not, and offer pi
#1049's "Node is installed on Windows but Settings says it isn't" was not a
bad heuristic — it was two probes for the same question, only one of them
hardened.

`apps/daemon/src/runtime/well_known_bin.rs` had already learned the lesson: a
bare `Command::new("node")` sees only the PATH this process inherited, so it
also searches `%ProgramFiles%\nodejs`, `%APPDATA%\npm`, the Homebrew prefixes,
and probes `.cmd`/`.bat` as well as `.exe`. Its own module doc names the
symptom exactly — "works perfectly in a terminal yet reports not installed in
the app". `apps/desktop/src/commands/deps.rs` never got any of it: it passed
the bare name for everything except opencode. So the same machine answered
"Node is fine" in onboarding (which asks amuxd) and "Node is missing" in
Settings.

Windows is where it bites hardest because nothing repairs that PATH there —
`fix_path_env` returns early ("GUI apps generally inherit the full PATH"),
which holds until the user installs Node after the app, or uses nvm-windows /
fnm / scoop, none of which touch the PATH the app inherited at launch.

So the lookup moves to a shared `teamclu-binpath` crate and both apps use it.
The daemon keeps its `well_known_bin` module as a re-export, so its ~10 call
sites are untouched and there is still exactly one implementation. The desktop
probe now also runs the child with `augmented_path()`: finding an npm shim is
not enough to start it, since its `#!/usr/bin/env node` needs node on the
child's PATH too. The brew pre-check inside `install_dependency` went through
the same resolver — it had the identical bug, and it guards an install.

Also from #1049: **pi is on the Dependencies page now.** A machine set up to
run pi previously saw only opencode there, listed as "required", with no way
to read pi's version or repair a broken install. It installs through
`amuxd install-pi`, and `run_amuxd_install_pi` is now shared with the first-run
wizard the way the opencode runner already was — one install path, two
callers. It is listed as optional, because which runtime matters is the user's
choice and this command has no view of it.

Not done, deliberately: pi gets no Update button. `UpdateButton` reads
`opencodeVersions` for both its "Up to date" badge and its "Update to {v}"
label, so listing pi as updatable would label it with opencode's version. An
installed-but-stale pi therefore still has to be repaired from onboarding.

The moved tests run in CI (`daemon-linux`, beside teamclu-runtime-env's) —
they would otherwise have run nowhere after the move — and deps.rs gains a
regression test for the tool-outside-PATH case that started all this.

Refs #1049
