+++ apps/daemon/Cargo.toml
@@ -36,6 +36,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 toml = "0.8"
 dirs = "6"
+teamclu-binpath = { path = "../../crates/teamclu-binpath" }
 
 # Utilities
 tracing = "0.1"
