+++ apps/desktop/src/commands/deps.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
 use std::process::Command;
 use tauri::{AppHandle, Emitter, Runtime};
 use tokio::io::{AsyncBufReadExt, BufReader};
@@ -39,31 +40,44 @@ pub struct DepInstallProgress {
     pub error: Option<String>,
 }
 
+/// A tool's own install directory, searched before the well-known ones.
+///
+/// Both installers hardcode a home-relative path that is on nobody's PATH.
+fn own_dirs(name: &str) -> Vec<PathBuf> {
+    let Some(home) = dirs::home_dir() else {
+        return Vec::new();
+    };
+    match name {
+        "opencode" => vec![home.join(".opencode").join("bin")],
+        "pi" => vec![home.join(".pi").join("bin")],
+        _ => Vec::new(),
+    }
+}
+
 /// Resolve the program to probe for a dependency's `--version` check.
 ///
-/// For most tools this is just the dependency name looked up on PATH. opencode
-/// is the exception: its official installer hardcodes `~/.opencode/bin`, which is
-/// NOT on the PATH a GUI app inherits when launched from Finder — so a bare
-/// `Command::new("opencode")` probe fails even though the binary is installed and
-/// amuxd (which resolves it by absolute path) runs it fine. Mirror amuxd's
-/// resolution order here: `~/.opencode/bin/opencode` (absolute) -> PATH fallback.
-/// See apps/daemon/src/opencode_install/mod.rs `resolve_binary`.
+/// This used to be `name.to_string()` for everything but opencode, and that is
+/// the whole of #1049's "Node is installed but Settings says it isn't": a bare
+/// `Command::new("node")` sees only the PATH this process inherited, and on
+/// Windows nothing repairs that PATH (`fix_path_env` returns early there), so a
+/// Node installed after the app — or by nvm-windows / fnm / scoop, which never
+/// touch the machine PATH — is invisible. amuxd answered correctly on the very
+/// same machine because it had already learned this and probes well-known
+/// directories; the two just never shared the lookup.
+///
+/// Now they do: `teamclu_binpath` is that shared lookup. `spawn_name` is the
+/// fallback rather than the bare name because Rust appends only `.exe` and
+/// never consults PATHEXT, so `Command::new("npm")` cannot start `npm.cmd`.
 fn probe_program(name: &str) -> String {
-    if name != "opencode" {
-        return name.to_string();
-    }
-    let bin = if cfg!(windows) {
-        "opencode.exe"
-    } else {
-        "opencode"
-    };
-    if let Some(home) = dirs::home_dir() {
-        let p = home.join(".opencode").join("bin").join(bin);
-        if p.exists() {
-            return p.to_string_lossy().to_string();
-        }
-    }
-    "opencode".to_string()
+    probe_program_in(name, &teamclu_binpath::search_dirs())
+}
+
+/// [`probe_program`] against an explicit directory list, so tests do not depend
+/// on what the machine running them happens to have installed.
+fn probe_program_in(name: &str, dirs: &[PathBuf]) -> String {
+    teamclu_binpath::find_with(name, &own_dirs(name), dirs)
+        .map(|p| p.to_string_lossy().to_string())
+        .unwrap_or_else(|| teamclu_binpath::spawn_name(name))
 }
 
 /// Check a single dependency by running `cmd --version` (or a variant).
@@ -77,8 +91,12 @@ fn check_single_dependency(
     affected_features: Vec<String>,
     priority: u8,
 ) -> DependencyInfo {
+    // Finding the file is not always enough to run it: an npm-installed shim
+    // starts with `#!/usr/bin/env node`, which needs node itself on the PATH of
+    // the child. Same augmentation amuxd spawns its tools with.
     let output = Command::new(probe_program(name))
         .no_window()
+        .env("PATH", teamclu_binpath::augmented_path())
         .args(version_args)
         .output();
 
@@ -200,6 +218,13 @@ fn get_install_commands_map(name: &str) -> Option<PlatformInstallCommands> {
             windows: "amuxd install-opencode".to_string(),
             linux: "amuxd install-opencode".to_string(),
         }),
+        // Same deal as opencode: display-only, with the real work done by the
+        // bundled sidecar in `install_dependency`.
+        "pi" => Some(PlatformInstallCommands {
+            macos: "amuxd install-pi".to_string(),
+            windows: "amuxd install-pi".to_string(),
+            linux: "amuxd install-pi".to_string(),
+        }),
         _ => None,
     }
 }
@@ -318,6 +343,22 @@ pub fn check_dependencies() -> Vec<DependencyInfo> {
         1,
     ));
 
+    // pi — optional, priority 1. The other runtime this app can install, and
+    // until #1049 the Dependencies page did not know it existed: a machine set
+    // up to run pi saw only opencode listed as "required", with no way to see
+    // pi's version or repair a broken install from here. Optional rather than
+    // required because which of the two matters is the user's runtime choice,
+    // which this command has no view of.
+    deps.push(check_single_dependency(
+        "pi",
+        &["--version"],
+        false,
+        "Agent runtime - the alternative to opencode for the local agent",
+        get_install_commands_map("pi").unwrap(),
+        vec!["Local Agent".to_string()],
+        1,
+    ));
+
     // Sort by priority (lower first)
     deps.sort_by_key(|d| d.priority);
     deps
@@ -336,10 +377,23 @@ pub async fn install_dependency<R: Runtime>(
     if name == "opencode" {
         return Ok(install_opencode_via_amuxd(&app, false).await);
     }
+    // pi likewise — `amuxd install-pi` is idempotent, so the same call installs
+    // it or lifts it to the version `pi.lock.json` pins.
+    if name == "pi" {
+        return Ok(install_pi_via_amuxd(&app).await);
+    }
 
     // On macOS, if the dependency requires brew and brew is not installed, install brew first
     if requires_brew(&name) {
-        let brew_check = Command::new("brew").no_window().arg("--version").output();
+        // Through the same resolver as every other probe: a Dock-launched app
+        // can hold a PATH without /opt/homebrew/bin in it, and a bare
+        // `Command::new("brew")` there reports "Homebrew is missing" on a
+        // machine that has it — then fails the install it was guarding.
+        let brew_check = Command::new(probe_program("brew"))
+            .no_window()
+            .env("PATH", teamclu_binpath::augmented_path())
+            .arg("--version")
+            .output();
         let brew_installed = matches!(brew_check, Ok(o) if o.status.success());
         if !brew_installed {
             let brew_result = run_install(&app, "brew").await;
@@ -432,6 +486,31 @@ async fn install_opencode_via_amuxd<R: Runtime>(app: &AppHandle<R>, force: bool)
     )
     .await;
     result.is_ok()
+}
+
+/// Install-or-upgrade pi via the bundled `amuxd install-pi`, bridging its
+/// progress onto `dep-install-progress` the same way opencode does.
+async fn install_pi_via_amuxd<R: Runtime>(app: &AppHandle<R>) -> bool {
+    let emit_app = app.clone();
+    crate::commands::setup::run_amuxd_install_pi(app, move |status, line, error| {
+        // amuxd emits "running"; the deps UI expects "installing".
+        let status = if status == "running" {
+            "installing"
+        } else {
+            status
+        };
+        let _ = emit_app.emit(
+            "dep-install-progress",
+            DepInstallProgress {
+                name: "pi".to_string(),
+                status: status.to_string(),
+                output_line: line,
+                error,
+            },
+        );
+    })
+    .await
+    .is_ok()
 }
 
 /// Execute the actual install command and stream output via events.
@@ -590,28 +669,72 @@ async fn run_install<R: Runtime>(app: &AppHandle<R>, name: &str) -> bool {
 mod tests {
     use super::*;
 
+    fn touch(dir: &std::path::Path, file: &str) -> PathBuf {
+        std::fs::create_dir_all(dir).unwrap();
+        let path = dir.join(file);
+        std::fs::write(&path, "").unwrap();
+        path
+    }
+
+    /// #1049: a Node that is installed but not on this process's PATH read as
+    /// "not installed" in Settings while onboarding — which asks amuxd, and
+    /// amuxd probes well-known directories — said it was fine.
     #[test]
-    fn probe_program_passes_through_non_opencode() {
-        assert_eq!(probe_program("gh"), "gh");
-        assert_eq!(probe_program("node"), "node");
+    fn a_tool_outside_path_is_still_found() {
+        let tmp = tempfile::tempdir().unwrap();
+        let dir = tmp.path().to_path_buf();
+        let node = touch(&dir, if cfg!(windows) { "node.exe" } else { "node" });
+        assert_eq!(probe_program_in("node", &[dir]), node.to_string_lossy());
     }
 
     #[test]
-    fn probe_program_resolves_opencode_absolute_when_present() {
-        // opencode's official installer targets ~/.opencode/bin; when that binary
-        // exists the probe must use its absolute path, not the bare PATH name a
-        // Finder-launched GUI can't resolve.
-        let bin = if cfg!(windows) {
-            "opencode.exe"
-        } else {
-            "opencode"
-        };
-        let expected = dirs::home_dir().map(|h| h.join(".opencode").join("bin").join(bin));
-        let got = probe_program("opencode");
-        match expected {
-            Some(p) if p.exists() => assert_eq!(got, p.to_string_lossy()),
-            // No installed binary in this environment: falls back to PATH lookup.
-            _ => assert_eq!(got, "opencode"),
+    fn a_tool_found_nowhere_falls_back_to_the_bare_name() {
+        let tmp = tempfile::tempdir().unwrap();
+        assert_eq!(probe_program_in("gh", &[tmp.path().to_path_buf()]), "gh");
+    }
+
+    /// Rust appends `.exe` and never consults PATHEXT, and npm ships no
+    /// `npm.exe` — only `npm.cmd`. The bare name can never start it.
+    #[test]
+    fn npm_falls_back_to_its_windows_shim_name() {
+        let tmp = tempfile::tempdir().unwrap();
+        let got = probe_program_in("npm", &[tmp.path().to_path_buf()]);
+        assert_eq!(got, if cfg!(windows) { "npm.cmd" } else { "npm" });
+    }
+
+    /// The runtimes' own installers hardcode a home-relative directory that is
+    /// on nobody's PATH, so it has to be searched before the shared list.
+    #[test]
+    fn a_runtimes_own_directory_beats_the_well_known_ones() {
+        let tmp = tempfile::tempdir().unwrap();
+        let well_known = tmp.path().join("well-known");
+        touch(&well_known, if cfg!(windows) { "pi.exe" } else { "pi" });
+
+        let got = probe_program_in("pi", std::slice::from_ref(&well_known));
+        let own = dirs::home_dir()
+            .map(|h| h.join(".pi").join("bin"))
+            .and_then(|dir| teamclu_binpath::find_with("pi", &[dir], &[]));
+        match own {
+            // A real pi install on this machine outranks the fixture.
+            Some(p) => assert_eq!(got, p.to_string_lossy()),
+            None => assert_eq!(
+                got,
+                well_known
+                    .join(if cfg!(windows) { "pi.exe" } else { "pi" })
+                    .to_string_lossy()
+            ),
+        }
+    }
+
+    /// Both runtimes must be offerable from the Dependencies page — pi was
+    /// missing entirely, so a pi machine had no way to see or repair it.
+    #[test]
+    fn both_agent_runtimes_have_an_install_path() {
+        for runtime in ["opencode", "pi"] {
+            assert!(
+                get_install_commands_map(runtime).is_some(),
+                "{runtime} needs install commands for the UI to render"
+            );
         }
     }
 }
