+++ crates/teamclu-binpath/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "teamclu-binpath"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+dirs = "6"
+
+[dev-dependencies]
+tempfile = "3"
